Better dependencies generation support in makefile Removed some compiler warnings Fixed some slowness problems in the planner screen
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@124 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/keys.cpp
@@ -54,7 +54,7 @@ static int keymap[128] =
 void prepare_keymap(const char *eng, const char *lng)
 {
 	assert(get_uformat() == U_UTF8);
-	assert(strlen(eng) == ustrlen(lng));
+	assert((int)strlen(eng) == (int)ustrlen(lng));
 
 	for (int i = 0; i < 128; i++) {
 		const char * p = strchr(eng, (char)i);
